Move CTransfer_SlidingInterface
