+++ SU2_CFD/src/transfer/physics/CTransfer_FlowTraction.cpp
@@ -506,107 +506,3 @@ void CTransfer_MixingPlaneInterface::SetAverageTurboGeoValues(CGeometry *donor_g
 
 }
 
-
-
-CTransfer_SlidingInterface::CTransfer_SlidingInterface(void) : CTransfer() {
-
-}
-
-CTransfer_SlidingInterface::CTransfer_SlidingInterface(unsigned short val_nVar, unsigned short val_nConst, CConfig *config) : CTransfer() {
-
-  rank = SU2_MPI::GetRank();
-  size = SU2_MPI::GetSize();
-
-  Physical_Constants = NULL;
-  Donor_Variable     = NULL;
-  Target_Variable    = NULL;
-
-  unsigned short iVar;
-
-  Physical_Constants = new su2double[val_nConst];
-  Donor_Variable     = new su2double[val_nVar];
-
-  Target_Variable    = new su2double[val_nVar+1];
-
-  valAggregated      = false;
-
-  nVar = val_nVar;
-
-  for (iVar = 0; iVar < nVar; iVar++) {
-    Donor_Variable[iVar]  = 0.0;
-    Target_Variable[iVar] = 0.0;
-  }
-
-  for (iVar = 0; iVar < val_nConst; iVar++) {
-    Physical_Constants[iVar] = 0.0;
-  }
-
-}
-
-CTransfer_SlidingInterface::~CTransfer_SlidingInterface(void) {
-
-}
-
-
-void CTransfer_SlidingInterface::GetPhysical_Constants(CSolver *donor_solution, CSolver *target_solution,
-                                                             CGeometry *donor_geometry, CGeometry *target_geometry,
-                             CConfig *donor_config, CConfig *target_config) {
-
-}
-
-void CTransfer_SlidingInterface::GetDonor_Variable(CSolver *donor_solution, CGeometry *donor_geometry, CConfig *donor_config,
-                                                unsigned long Marker_Donor, unsigned long Vertex_Donor, unsigned long Point_Donor) {
-
-  unsigned short iVar, nDonorVar;
-  nDonorVar = donor_solution->GetnPrimVar();
-
-  /*---  the number of primitive variables is set to two by default for the turbulent solver ---*/
-  bool turbulent = (nDonorVar == 2) ;
-
-  if (turbulent){
-
-    /*---  for turbulent solver retrieve solution and set it as the donor variable ---*/
-    Donor_Variable[0] = donor_solution->node[Point_Donor]->GetSolution(0);
-    Donor_Variable[1] = donor_solution->node[Point_Donor]->GetSolution(1);
-
-  } else{
-
-    /*---  Retrieve primitive variables and set them as the donor variables ---*/
-    for (iVar = 0; iVar < nDonorVar; iVar++)
-    Donor_Variable[iVar] = donor_solution->node[Point_Donor]->GetPrimitive(iVar);
-
-  }
-}
-
-void CTransfer_SlidingInterface::InitializeTarget_Variable(CSolver *target_solution, unsigned long Marker_Target,
-                          unsigned long Vertex_Target, unsigned short nDonorPoints) {
-
-  target_solution->SetnSlidingStates(Marker_Target, Vertex_Target, nDonorPoints); // This is to allocate
-  target_solution->SetSlidingStateStructure(Marker_Target, Vertex_Target);
-  target_solution->SetnSlidingStates(Marker_Target, Vertex_Target, 0); // Reset counter to 0
-
-}
-
-void CTransfer_SlidingInterface::RecoverTarget_Variable(long indexPoint_iVertex, su2double *Buffer_Bcast_Variables,
-                                                        su2double donorCoeff){
-  for (unsigned short iVar = 0; iVar < nVar; iVar++)
-      Target_Variable[iVar] = Buffer_Bcast_Variables[ indexPoint_iVertex*nVar + iVar ];
-
-  Target_Variable[nVar] = donorCoeff;
-}
-
-void CTransfer_SlidingInterface::SetTarget_Variable(CSolver *target_solution, CGeometry *target_geometry,
-                          CConfig *target_config, unsigned long Marker_Target,
-                          unsigned long Vertex_Target, unsigned long Point_Target) {
-
-  unsigned short iVar, iDonorVertex, nTargetVar;
-  nTargetVar = target_solution->GetnPrimVar();
-  /*--- Set the Sliding solution with the value of the Target Variable ---*/
-
-  iDonorVertex = target_solution->GetnSlidingStates(Marker_Target, Vertex_Target);
-
-  for (iVar = 0; iVar < nTargetVar+1; iVar++)
-    target_solution->SetSlidingState(Marker_Target, Vertex_Target, iVar, iDonorVertex, Target_Variable[iVar]);
-
-  target_solution->SetnSlidingStates( Marker_Target, Vertex_Target, iDonorVertex + 1 );
-}
